libs/service-commons/pyproject.toml
[project]
name = "service-commons"
version = "0.1.0"
description = "Shared infrastructure utilities for artwork-matcher services"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/service_commons"]
include = ["src/service_commons/py.typed"]
